+++ zsocket.go
@@ -106,17 +106,17 @@ func errnoErr(e syscall.Errno) error {
 	return e
 }
 
-func copyFx(dst, src []byte, len uint32) {
+func copyFx(dst, src []byte, len uint16) {
 	copy(dst, src)
 }
 
 type IZSocket interface {
 	MaxPackets() uint32
-	MaxPacketSize() uint32
+	MaxPacketSize() uint16
 	WrittenPackets() uint32
 	Listen(fx func(*nettypes.Frame, uint16))
-	WriteToBuffer(buf []byte, l uint32) (int32, error)
-	CopyToBuffer(buf []byte, l uint32, copyFx func(dst, src []byte, l uint32)) (int32, error)
+	WriteToBuffer(buf []byte, l uint16) (int32, error)
+	CopyToBuffer(buf []byte, l uint16, copyFx func(dst, src []byte, l uint16)) (int32, error)
 	FlushFrames() (uint, error, []error)
 	Close() error
 }
@@ -128,12 +128,12 @@ type ZSocket struct {
 	raw            []byte
 	listening      int32
 	frameNum       uint32
-	frameSize      uint32
+	frameSize      uint16
 	rxEnabled      bool
 	rxFrames       []*ringFrame
 	txEnabled      bool
 	txLossDisabled bool
-	txFrameSize    uint32
+	txFrameSize    uint16
 	txIndex        int32
 	txWritten      uint32
 	txWrittenIndex int32
@@ -228,7 +228,7 @@ func NewZSocket(ethIndex, options, blockNum, frameOrder, framesInBlock int, ethT
 	}
 	zs.raw = bs
 	zs.frameNum = uint32(req.frameNum)
-	zs.frameSize = uint32(req.frameSize)
+	zs.frameSize = uint16(req.frameSize)
 	i := 0
 	frLoc := 0
 	if zs.rxEnabled {
@@ -240,7 +240,7 @@ func NewZSocket(ethIndex, options, blockNum, frameOrder, framesInBlock int, ethT
 		}
 	}
 	if zs.txEnabled {
-		zs.txFrameSize = zs.frameSize - uint32(_TX_START)
+		zs.txFrameSize = zs.frameSize - uint16(_TX_START)
 		zs.txWritten = 0
 		zs.txWrittenIndex = -1
 		for t := 0; t < int(zs.frameNum); t, i = t+1, i+1 {
@@ -260,7 +260,7 @@ func (zs *ZSocket) MaxPackets() uint32 {
 }
 
 // Returns the frame size in bytes
-func (zs *ZSocket) MaxPacketSize() uint32 {
+func (zs *ZSocket) MaxPacketSize() uint16 {
 	return zs.txFrameSize
 }
 
@@ -303,12 +303,12 @@ func (zs *ZSocket) Listen(fx func(*nettypes.Frame, uint16)) error {
 
 // WriteToBuffer writes a raw frame in bytes to the TX ring buffer.
 // The length of the frame must be specified.
-func (zs *ZSocket) WriteToBuffer(buf []byte, l uint32) (int32, error) {
+func (zs *ZSocket) WriteToBuffer(buf []byte, l uint16) (int32, error) {
 	if l > zs.txFrameSize {
 		return -1, fmt.Errorf("the length of the write exceeds the size of the TX frame")
 	}
 	if l < 0 {
-		return zs.CopyToBuffer(buf, uint32(len(buf)), copyFx)
+		return zs.CopyToBuffer(buf, uint16(len(buf)), copyFx)
 	}
 	return zs.CopyToBuffer(buf[:l], l, copyFx)
 }
@@ -317,7 +317,7 @@ func (zs *ZSocket) WriteToBuffer(buf []byte, l uint32) (int32, error) {
 // ring buffer. However, it can take a function argument, that will
 // be passed the raw TX byes so that custom logic can be applied
 // to copying the frame (for example, encrypting the frame).
-func (zs *ZSocket) CopyToBuffer(buf []byte, l uint32, copyFx func(dst, src []byte, l uint32)) (int32, error) {
+func (zs *ZSocket) CopyToBuffer(buf []byte, l uint16, copyFx func(dst, src []byte, l uint16)) (int32, error) {
 	if !zs.txEnabled {
 		return -1, fmt.Errorf("the TX ring is not enabled on this socket")
 	}
@@ -501,12 +501,12 @@ func (rf *ringFrame) tpLen() uint16 {
 	return uint16(inet.Int(rf.raw[_TP_LEN_START:_TP_LEN_STOP]))
 }
 
-func (rf *ringFrame) setTpLen(v uint32) {
-	inet.PutInt(rf.raw[_TP_LEN_START:_TP_LEN_STOP], v)
+func (rf *ringFrame) setTpLen(v uint16) {
+	inet.PutInt(rf.raw[_TP_LEN_START:_TP_LEN_STOP], uint32(v))
 }
 
-func (rf *ringFrame) setTpSnapLen(v uint32) {
-	inet.PutInt(rf.raw[_TP_SNAPLEN_START:_TP_SNAPLEN_STOP], v)
+func (rf *ringFrame) setTpSnapLen(v uint16) {
+	inet.PutInt(rf.raw[_TP_SNAPLEN_START:_TP_SNAPLEN_STOP], uint32(v))
 }
 
 func (rf *ringFrame) rxReady() bool {
